Fix -Wconversion warnings: add explicit casts for size_t/float conversions in LLM modules
- llama_wrapper.cpp: Add explicit casts to size_t for reserve() calls with int parameters
  - Lines 1387, 2853, 3135: Cast max_tokens (int) to size_t for vector reserve()
- lora_training_service.cpp: Add explicit casts for float division operations
  - Line 85: Cast predictions.size() to float in MSE loss computation
  - Line 95: Cast predictions.size() to float for scale factor calculation

These changes fix -Wconversion warnings that result from implicit narrowing
conversions or type mismatches between signed/unsigned integer types and
floating-point types during arithmetic operations.

+++ src/llm/lora_framework/lora_training_service.cpp
@@ -82,7 +82,7 @@ float compute_mse_loss(const Tensor& predictions, const Tensor& targets) {
         sum += diff * diff;
     }
     
-    return sum / predictions.size();
+    return sum / static_cast<float>(predictions.size());
 }
 
 // Compute gradient of MSE loss w.r.t. predictions
@@ -92,7 +92,7 @@ Tensor compute_mse_gradient(const Tensor& predictions, const Tensor& targets) {
     }
     
     Tensor grad(predictions.shape());
-    float scale = 2.0f / predictions.size();
+    float scale = 2.0f / static_cast<float>(predictions.size());
     
     for (size_t i = 0; i < predictions.size(); ++i) {
         grad[i] = scale * (predictions[i] - targets[i]);
